Welcome to Crystal Adventure, Volume 1 Episode 1: Battle of Eindhoven -- Ik Spreek Geen Nederlands
You find yourself in a small clearing.  To the north you see a castle.
To the east and south is a dark and impassable forest.  A windmill can
be seen in the distance to the west.

A Dutch dwarf stands nearby holding a picnic basket and humming
tunelessly to himself.

What would you like to do next?
> attack dwarf with soup spoon

You attack the dwarf with a spoon.  The dwarf suffers 34 damage points.
The dwarf retaliates and attacks you with a salad fork.

You suffer 61 damage points.
> attack dwarf with butter knife

You attack the dwarf with a knife.  The dwarf suffers 83 damage points
and dies.