This commit renames INTEL output directory to X86.
As I was complained to that INTEL is a brand name
and not the only X86 processor manufacturer.

* IMPORTANT *

Once you update people should delete the contents
of your "CS\out" directories and do a full and clean
rebuild of the Crystal Space code.

This update has been tested on VC and with MingW,
but should be Ok for all other GCC builds / ports :-)