Fixed so will compile on MSVC.
This fix has been tested on MSVC, MingW (GCC)
and the GCC 2.96 from Linux Mandrake 8 and
works fine. This all could be altered at a lter date
when the plugin bacomes active, but for testing
purposes on MSVC it is adequate.

Also removed:

/* Scanner skeleton version:
 * $Header: /cvsroot/crystal/CS/plugins/aws/sllex.cpp,v 1.6 2001/06/15 22:27:56 sunshine Exp $
 */

which Erics terminal merged this into the file.