- Updated FreeFont 2 plugin makefile for successful compilation with MSVC.
- Upated msvc_libs adding Freetype 2.0.8.

Note:

This version of FreeType2 is a mod version from the official Freetype 2.0.8 distro.
This is due to the poor quality of MSVC project file maintenance where the
output filename does not match the distro release number (2.0.6 is outputted
and not 2.0.8 as it should be. Fixed by me).

Also the Freetype project uses an absolutley stupid method for #include which
MSVC really does not like. In certain circumstances I have worked around this
in the Free Font 2 MSVC project file and in some cases by hardcoded includes.

- Updated docs/links to reflect the locations of new msvc_libs.

Note:

This plugin has not yet added to the default MSVC build process, though
it compiles without error. Until Norman Kramer gives me the nod to say it is
ready, that is how it will stay. :-) If you just can't wait uncomment:

#PLUGINS.DYNAMIC += font/server/freefnt2

in "CS\mk\user.mak"

and rebuild the MSVC project files via the documented method.

Regards

Philip Wyett