- reworked the Bullet plugin and implemented many of the missing functionalities - added missing methods about capsule geometry in dynamics.h - improved the API documentation of dynamics.h - changed deprecated method in odedynam.cpp
Note that the changes on the Bullet plugin have been made with Bullet 2.75, it has NOT been tested with Bullet 2.68. Bumping the version of the 
bullet's dependency is probably a good idea since many fixes were made in the library between 2.68 and 2.75.

Note also that the configure check for Bullet is still unclear. The bullet.pc.in specifies that the code should be linked with '-lbulletdynamics 
-lbulletcollision -lbulletmath', while the names of the generated libraries should imply '-lBulletDynamics -lBulletCollision -lLinearMath' (well, at 
least with my config on my Ubuntu Hardy 64bits).