Changed the coordinate system of the transforms manipulated by the animation nodes to bind space instead of bone space.
The reason of this is that the neutral position of the skeleton of an animesh should correspond to an empty transform of all the bones. This is true in bind space, not in bone space.
Not doing this may generate some problems when blending partially an animation. For example, in r34293 with avatartest and the Frankie scene, using the + and - keys to change the speed of the walk, the eyes will not be blend correctly.