Omit failed checks from summary of features foreign to current platform.
Some notable configure checks are guaranteed or very likely to fail on
specific platforms, thus they become annoying noise when reported as
"missing" in the configure summary report.  For example, the
"cs-winlibs" check is worthy of reporting in the configure summary on
Windows, but will always fail on Unix, and thus becomes mere noise in
the Unix summary. Likewise, although the GTK check may succeed on
Windows and should be reported successful in the summary, in most cases
it will fail and such expected failure should be suppressed.

Consequently, add support to CS_NOTABLE() and CS_SUMMARIZE() in
summary.m4 for suppressing specified classes of failed checks from the
configure summary report.