vararg: Add CS::va_call() utility.
The llvm-gcc compiler shipped by Apple with Xcode is incompletely
implemented.  In particular, when it encounters a vararg virtual method
in a virtual base class, it aborts with error "generic thunk code fails
for method foo::bar() which uses '...'". This is particularly
problematic in CS since virtual inheritance is employed heavily by the
SCF facility, and it is not entirely uncommon for vararg methods to be
declared in SCF interfaces.  There are at least two possible ways to
work around this shortcoming.

The first work-around is to #ifdef-out the offending vararg methods and
rely only on their va_list counterparts. This is somewhat ugly since it
puts burden on the callers of those methods, however, it is fairly
non-invasive and results in relatively little code churn. Another
downside is that it introduces API breakage in the interfaces containing
vararg methods, however, this breakage is essentially forced upon us by
Apple rather than being deliberate on our part. All projects attempting
to build with Apple's broken compiler are faced with the same choice, so
it is not as if this API change places any additional burden on foreign
projects than that already caused by Apple.

The second work-around is to abandon multiple-inheritance for
implementing SCF interfaces in favor of implementation via embedding.
This is how SCF was designed to work originally when compiler support
for C++ multiple inheritance was not widespread. The benefit of this
approach is that it retains perfect API compatibility for clients, but
at the expense of significant work re-engineering each class
implementing an SCF interface with vararg methods. Moreover, it is
likely that the modern template-based version of SCF does not even
support interface implementation via embedding, so it may not even be an
option.  Finally, a number of years ago, a great deal of effort was
undertaken to purge CS of embedded interfaces, hence it would be a step
backward to re-introduce it.

In the interest of minimizing code churn, and for its general ease of
implementation, at this time, the first work-around seems the better
choice.  This means that foreign clients, as well as CS itself, which
wish to support OSX will need to avoid invoking vararg methods in SCF
interfaces.

Introduce CS::va_call() and CS::va_callv() functions to help ease the
transition from invocation of vararg methods to their va_list
counterparts.  va_call() is for invoking methods which return a value,
whereas va_callv() is for methods declared void.  To employ these
convenience functions, consider the following interface:

struct iFoo {
  void Print(char const* fmt, ...) = 0;
  void PrintV(char const* fmt, va_list) = 0;
};

The normal way to print via iFoo is:

iFoo* foo = ...;
foo->Print("%d %s", 42, "forty-two");

However, clients wishing to support OSX should instead invoke the
va_list counterpart via CS::va_callv(), which provides a similar
interface with relatively little extra pain:

iFoo* foo = ...;
CS::va_callv(&iFoo::PrintV, foo, "%d %s", 42, "forty-two");