docs: macosx: Retire advice to place 3rd-party libraries in CS/libs.
The advice to build 3rd-party prerequisite libraries directly in
CS/libs, when lacking superuser privilege, is from ancient times when it
was the approach recommended with the original CS build system.  These
days it is better to install 3rd-party libraries via a package manager
or, if by hand, to install them in well-known locations outside the CS
tree.  Moreover, the functionality of CS configure to search for
3rd-party libraries in CS/libs is deprecated and may be removed in the
future.