bruteblock: Resolve "deprecated method" warnings.
Methods iTerrainObjectState::SetMaterialMap() and SetMaterialAlphaMaps()
are correctly tagged as deprecated in the interface file.  Applying
CS_DEPRECATED_METHOD again in the implementation file is pointless and
results unnecessary "deprecated method" compilation warnings which are
pure noise (that is, they do not, in this case, notify the developer of
a potential problem requiring attention).