swig: Suppress warning about unused label in generated code.
Modern SWIG emits a 'fail:' label in every C function it generates but
often does not emit a corresponding 'goto', hence the compiler warns
about the unused label. This warning is pure noise since it arises from
generated code over which we have no control and since it is not
indicative of any problem on our side. Therefore, suppress it.