Refactored PostEffect
Layers now have a user shader variable context and a default context for defaul shader parameters.
Added 4 functions to setup (get/set) the input and output of the PostEffect. The first one is the input texture and the second is the target where the posteffect will be drawn.

When drawing a posteffect now there is 3 options:
Draw to target texture, or to the screen and "none" where the posteffect is drawn in the texture allocated internally.

Added the function "LoadFromFile" and "Construct" in iPostEffect, the second one is responsible to completely setup the effect making it ready to use.

Added DependencySolver class to optimize texture usage given a graph of layers.

PostEffectManager now handles the texture allocation/sharing for post effects.

Added PriorityShaderVariableContext class where can be added multiples parent sv contexts with different priorities


TODO: fix hdr code to properly work with these changes.